fix(file): rename request/capa.go and request/capx.go
rename capx.go to capa.go and capa.go to capx.go
